[movies] init: Add gradle dexcount plugin
This plugin makes it easy for a project to observe its dexcount grow as
the project grows and new libraries are added to it. Also, it provides a
graphical representation of all the project's dependencies and how much
those add to the overall dexcount.

In addition, with this configuration the project build will fail if the
dexcount exceeds the limit of 60K methods. This is so the project
doesn't exceed the limit of 65K and thus becomes a multidex project,
which is not an ideal situation to be in.