[movies] init: Add architecture components lifecycle extensions
This is one of the most useful components of the architecture component
family that Google announced during GoogleIO17. Amongst other utilities
it provides, for this specific case, it provides us with the
ProcessLifecycleOwner, out of which the app lifecycle can be obtained
and used as an observer for the application class so that the
architecture components @OnLifecycleEvent annotations can be utilised.

For this specific scenario it helps add logging on the application
level for all app level lifecycle events. This, ultimately helps
developers know in which state the app is, at any specific moment, and
easily debug application lifecycle related flows.