[movies] feature: Add movie year picker
This picker will allow us to filter the movies based on their release
year. Because of the fact that Android by default provides a date
picker, without any possibility to show only the year, or month, or any
other combination, the community created a utility library that does
exactly that. Thus, this commit adds this very thin library to quickly
allow for such functionality.

Android Pickers:
https://developer.android.com/guide/topics/ui/controls/pickers

Community MonthAndYearPicker:
https://github.com/premkumarroyal/MonthAndYearPicker