[movies] migration: Migrate to gradle kotlin dsl - versions plugin
Note that the migration is not complete on this plugin since the
dependency_update.gradle file is still used through the apply(...)
mechanism.

The apply cannot be replaced with the actual dependency_update.gradle
logic since the plugin itself is written in Groovy language and
importing the plugins "dependencyUpdates" def does not work.