[movies] migration: Migrate to gradle kotlin dsl - android plugin
Note that the migration is not complete on this plugin since the
android.gradle and lint.gradle files are still used through the
apply(...) mechanism.

The apply cannot be replaced with the actual android.gradle and
lint.gradle logic since there are two distinct plugins for android, one
for the "data" module, through the android library plugin, and the
other one for the "presentation" module, through the android
application plugin.