[movies] build: Add plugin build logs
This logs will help identify which plugin is running for which
configuration and further more help us identify those plugins that are
never triggered due to work configuration.

For example, having done the above, it was evident that the plugin
block for KotlinPlatformCommonPlugin::class was never triggered. Thus,
all the configuration within is unused, which is wrong, as this
configuration needs to work for the build to do what is expected of it.