[movies] migration: Migrate to gradle kotlin dsl - android plugin conf.
Split android (BasePlugin) plugin configuration into two, the android
library plugin (LibraryPlugin) and the android application plugin
(AppPlugin). This is so to distinguish when each of those plugins run
for which configuration.

Furthermore, this might give us the possibility to convert the Groovy
related android.gradle and lint.gradle into Gradle Kotlin DSL.