[movies] refactor: Make expect that to more idiomatic
In addition to "expectThat(...).isXyz(...)", Strict support a more
idiomatic way of using lambdas to execute assertions on multiple
subjects at once. Thus, this change makes sure to replace all
occurrences of expectThat(...) with the more idiomatic
"expect { that(...).isXyz(...) }" style.

As per Strict's statement, the advantage of using this style of doing
assertions is that it supports multiple subject at once (if/when
they are/will be needed).

Thus, from now on "expect { that(...).isXyz(...) }" will become the
default when or using Strict's assertions.