[movies] migration: Migrate to junit5 and spek, dependencies junit4
This configuration makes sure that both Spek and JUnit4 tests can run
simultaneously. JUnit4 tests are still required for Robolectric tests,
thus this is inevitable until a good solution is found to write
Robolectric tests with Spek and the "RobolectricTestRunner" runner,
which is based on JUnit4.

Currently the JUnit4 Vintage configuration is only required for Android
specific modules, and more specific for the "presentation" module,
which has extra JUnit4 dependencies configured for its Robolectric
tests (which is not required for the "data" module). However, the pure
Kotlin specific modules (currently the "domain" module) do not require
this configuration and thus they utilise just the Spek engine.