[movies] cleanup: Remove gradle versions plugin
This plugin does not work with Gradle 6 and Android Studio 4. At a
later point in time another plugin should take its place, one that
works.

Also, with this change the last groovy gradle file was removed as well,
the 'dependency_updates.gradle'. Thus, with that change this project is
fully on Kotlin Gradle DSL.