[movies] test: Replace run blocking with suggested run blocking test
See the below note from 'Using Kotlin Coroutines in your Android App'
codelab, step 10 (Testing coroutines directly):
'Important: The function runBlockingTest will always block the caller,
just like a regular function call. The coroutine will run synchronously
on the same thread. You should avoid runBlocking and runBlockingTest in
your application code and prefer launch which returns immediately.
- runBlockingTest: should only be used from tests as it executes
coroutines in a test-controlled manner, while
- runBlocking can be used to provide blocking interfaces to coroutines.'

Also, see this additional note from the same step:
'runBlockingTest relies upon TestCoroutineDispatcher to control
coroutines. As a result, it's a good idea to inject a
TestCoroutineDispatcher or TestCoroutineScope when using
runBlockingTest. This has the effect of making the coroutines single
threaded and offers the ability to explicitly control all coroutines in
tests. If you don't want to change the behavior of coroutines – for
example in an integration test – you can instead use runBlocking with
the default implementations of all dispatchers.'

Finally, please see this extra note at the end from the same step:
'runBlockingTest is experimental, and currently has a bug that makes it
fail the test if a coroutine switches to a dispatcher that executes a
coroutine on another thread. The final stable is not expected to have
this bug.'