[movies] build: Add new variant configuration for lib and app modules
This configuration is responsible to filter out unnecessary variants,
based on the new 'ignored_variants' configuration, which is located
within the 'local.properties' file. This is done in order to speed up
build times and complexity during development, thus making developers
happier since in most cases developers usually only care about the
'debug' build type (and a single build flavor, if that is even
available).

This new configuration is located within the 'local.properties' file in
order to have every developer specify it locally and not have it under
version control. If no configuration is specified, then the default
configuration will be applied, which is all build types and flavors.

For developers to apply this configuration they have to first locate
the 'local.properties' file and if it doesn't exist already, then
create it. Then, they need to add the 'ignored_variants' property and
assign that to a value. This value can be a single variant value, such
as the 'release' built type or a comma separated list of variant
values, such as 'debug,release,etc'. Any variant that goes in there
will be then ignored by the build.