[movies-17] migration: Migrate to result
As part of this migration the below occurred:
1) The 'Result' generic domain model was introduced. This is the main
sealed class that holds all the result related functionality, which has
both the 'Success' and 'Error' subtypes. Also an extra 'asResult(...)'
extension function was introduced, which helps remove result related
boilerplate code.
2) The 'Errors' generic domain models were introduced. This is the main
file that holds all the error related classes, those are the ones that
are being converted from exceptions. Also an extra 'toError(...)'
extension function was introduced, which is used alongside the
'asResult(...)' extension function and is responsible to automatically
convert exceptions to errors for all the exceptions that are marked as
mappable.
3) The 'Exceptions' generic network models were introduced. This is the
main file that holds all network mappable exceptions, those are the
ones that are being converted to errors. Also an extra
'withException(...)' extension function  was introduced, which helps
remove network exception related boilerplate code.