[movies-18] feature: Create coverage plugin
This is a custom plugin that is responsible to read the 'jacoco.xml'
file that is being generated when the 'jacoco' task is triggered. The
actual task is named as 'coverage'.

Afterwards, and based on the 'coverage.properties' that should exist
per module, the results are being processed and compared with each
other. A success or failure output is shown depending on the comparison
results.

The 'coverage' task gets to be successful if the actual coverage
results are higher or at least equal to the expected coverage
percentage (per coverage type), which is defined within the
'coverage.properties' file. This means that if the numbers are lower,
the build will fail as the coverage for that specific module has
dropped from the expected.

Moreover, the newly created 'coverage' task depends on the 'jacoco'
task and as such cannot run if 'jacoco' is not run as well. In case
'jacoco' was not triggered beforehand, the 'coverage' task will trigger
it automatically before progressing with its own task.