[movies] build: Move build configuration classes to new config package
This is done in order to structure even further the configuration
classes and files. The only exception was the 'Plugin' related
constants, which could not be moved. As thus, a newly created object
class, the 'Plugins' was created, located outside the package structure
in order to make it possible for the 'buildscript { ... }' block to
access these constants.

Also, as part of this change the 'Reps' object class and its usage was
removed since it was actually not necessary for some time now.