[movies-19] test: Add rest client integration test to network module
This is the first integration test added to the network module. Its
main job is to provide confident that the mapping from raw responses to
domain models are correct.

A 'movies_page.json' file is provided with with real world example of
how such a response is provided when the MovieDB is being queried for
a list of movies.