[movies-20] test: Add integration tests for first, second and last pages
In addition to testing the first page, this change adds testing for the
second and last page. The movies pages json files were updated to
reflect that change and to make it so that the tests now replicate the
real world scenario end-to-end, from start to finish.

Also, this change increases our test coverage since it takes into
consideration the transformation of the last page to a 'MoviesPage'
domain model that has no 'nextPage'.