[movies-21] test: Remove toolbar from jacoco coverage excludes config
Since both toolbars are now covered by test, this filter is no longer
necessary. Only activity related code is left untested, but this will
be covered with Espresso tests later on. 'module', 'binding' and
'glide' related code is left untested on purpose and will not be tested
at all since there is no real benefit in tested this piece of code.