[movies-22] refactor: Replace val/var to fun modules
This way the modules can be further parameterized if need be. This will
be later on required for creating a mechanism to pass the 'baseUrl'
string parameter from the 'App' class for the 'NetworkModule' related
configuration. Thus, by having this mechanism, another
'InstrumentedTestApp' class can be created, which will extend the 'App'
class and override the 'baseUrl' with that of the 'MockWebServer'
related url. This way, the later on Espresso test will be able to
benefit from fake responses, thus making testing totally deterministic.