[movies-22] test: Change the web mock server port to a random port
This is done in order to avoid 'BindException' being triggered during
tests on machines that have already bind a socket to a local address
and port such as http://localhost:8080, which is pretty easy to do
since 8080 is a very common port to use.

Thus, this 54424 port number is now used instead of 8080 to make it
much more difficult to encounter the 'BindException' again in the
future.