[movies-22] test: Add first app e2e instrumented test
Note the fact that this test is not using the robot pattern yet. This
was done on purpose so that during subsequent commits it can be shown
how Espresso tests can be further improved so that everyone can easily
read and write them without the need for complicated helper functions
and extra setup.