[movies-22] test: Change filter and picker month from april to may
This is done because today is the first of May 2020 and as such the
picker default to the current month, which is May (instead of April).
Thus, and unless the tests are updated so that they specifically select
April as a month, those tests will fail. This is a quick fix with
attached TODOs so that a permanent fix is introduced in the future.