[movies-24] refactor: Improve load data or restore functionality
Previously, when a configuration change was occurring, the adapter was
responsible to identify whether that should be a load or a reload of
items. With this change, this logic now resides within the activity,
which holds a reference to a 'RELOAD_ITEMS' boolean that is saved
within the instant state.

This way the whole logic becomes more simple and will be subsequently
work with the new MVI architecture quite well.