[movies-24] feature: Add single live event
This class is copy-pasted from Google's architecture samples todoapp
blueprint repository. It will be subsequently used with the new MVI
architecture for side effects.

As the Javadoc suggest, it is a lifecycle-aware observable that sends
only new updates after subscription, used for events like navigation
and Snackbar messages.

This avoids a common problem with events: on configuration change (like
rotation) an update can be emitted if the observer is active. This
'LiveData' only calls the observable if there's an explicit call to
'setValue()' or 'call()'.

Note: Only one observer is going to be notified of changes.