[movies-24] feature: Add base mvi activity and view model classes
These mvi activity and view model classes are going to be used as base
classes for all the screens that are build around with MVI architecture.

The main components of this MVI architecture are:
- Intent: This is the event that the screen passes down for processing.
- Action: This is the action that is produced either by the events send
for processing or other triggers. The action, along with the state (or
side effect) is being fed to the reducer, which produces a new state
(or side effect).
- State: This is the main state that the screen is observing, which is
produced by the events send for processing or other triggers. The
state, along with the action is being fed to the reducer, which
produces a new state.
- Side Effect: This is an optional side effect state that the screen is
observing, which is produced by the events send for processing or other
triggers. An action is being fed to the reducer, which produces a side
effect.
- Reducer: This is the mechanism that produces new a state or side
effect, given an action an a previous state (for state). For side
effects, only an action is required.