[movies] fix: Improve infinite scroll listener
With this little tweak 'listener.loadMore()' will not trigger twice or
unnecessary when it is not required. Only when a non-zero vertical
scroll will now trigger the 'onScrolled(...)' logic on the
'InfiniteScrollListener'.

Actually, this fix improves the instrumentation test since the second
page is no longer required to be mocked (on all scenarios). Thus, all
the instrumentation test related code that had to do with loading and
testing the second page is now gone.