[movies] feature: Enhance movies mvi with idle state
This step is done as a preparation step for the 'Navigation Component'
functionality. Since the 'MoviesFragment', which will replace the
'MoviesActivity', will go through the 'onCreateView(...)' phase every
time the user navigates to and back from the 'MovieDetailsFragment',
this extra functionality will make sure that the all the state (toolbar
included) will be set from scratch.