[movies-25] feature: Add navigation component
As part of this commit the below occurred:
1) The navigation component plugin and dependencies were added,
2) The navigation component plugin was added to the root project,
3) The navigation component dependencies were added on the 'app',
'movies' and 'movie-details' modules since this is were the navigation
occurs. The 'app' module will host the 'AppActivity', which will become
the single activity of the project. The 'movies' and 'movie-details'
modules will host the destination fragments.
4) The navigation component plugin was applies on the 'movie-details'
module since this is the only module that will be using 'safe-args'.