[movies-25] feature: Add app activity with app navigation and toolbar
This will become the single activity of the project.

Also, as part of this commit and the 'app_activity' layout, a
'DrawerLayout' is added to the project, but which will be left empty on
purpose for now and be filled with functionality as part of another
future item of this repo.