[movies-25] feature: Replace app activity with splash activity
Now the app activity will become and entry point to the app. It will be
then responsible to navigate the whichever screen will be defined
within the 'app_navigation' graph. All that will be done in a
subsequent commit.

NOTE: As it is now, the app only shows a splash screen and then closes.