[movies] fix: Update tools content with full package name
Since the auto-manifest plugin created new package names for the
Android modules, those do not match anymore. Thus, the full package
name is required for it to work. Another optimization would have been
to update the underlying package structure of the Android modules as
well, but this is left as potential future work.