[movies] refactor: Improve pagination data domain model logic
Since MVI is now in place, there is no reason for keeping any logic
within this data class. All this is now moved to the respective
reducer, which will be now responsible to do the 'addPage(...)' logic.

Thus, this little improvement makes 'PaginationData' totally immutable
and as such it is effectively no longer a shady data class.