[movies] fix: Dismiss the snackbar when leaving the screen
This bug was identified when in the movie details screen. In that
screen when a loading error occurs and the snackbar is shown, if the
user goes back to the previous screen the snackbar continues to show
(but it shouldn't have). Thus, if the user then click on the 'RETRY'
action the app will crash.

This commit fixes this issue and a potential similar issue on the
movies screen.