[movies] refactor: Add binding generic to base activity and fragment
This way the client that extend and utilise the base and mvi screen
will be forced to supply a view binding to be able to build their
screens. It is like a contract that is required from now on to make the
development of all screens within this project view binding only.