[movies] update: Update android studio to canary 4.2.0-alpha01 version
Also, as part of this update, Gradle is updated to '6.5' from its
'milestone-1' equivalent since the previous 'Could not determine the
dependencies of task ':packageExternalRelease'.' problem is now gone
when using '4.2.0-alpha01'.

Actually, the '4.2.0-alpha01' version of Android Studio cannot be used
with a Gradle version less than '6.5', thus this change is totally
mandatory.