[movies] lib: Add leak canary plumber
LeakCanary Plumber performs hacks at runtime to fix some of the leaks
caused by a known bug in 3rd party code that this codebase have no
control over.

Since the 'leakcanary-android' dependency is usually added as a
'debugImplementation' dependency, the 'plumber-android' is transitively
added only in debug builds, so it will not fix leaks in the release
builds. Thus, the dependency is directly added as 'implementation' to
get these fixes in release builds as well.