[movies] test: Add custom robolectric test runner
This runner will replace all the individual Robolectric configuration
per Robolectric test, thus making the creation of Robolectric tests
more straightforward.

Also, as part of this enhancement the explicit setting up of the SDK
level was added, which points to Android Q (29). This is done so that
updating the target/compile SDK becomes as smooth as possible, without
the need to update Robolectric as well. For example, while updating to
Android R (30), Robolectric will fail complaining that the API level 30
is not available. However, with this configuration it will not since it
now explicitly point to API level 29 anyway.

Also, note that with that change another problem bubbled up. Android
Studio started throwing an 'java.lang.UnsupportedOperationException'
stating that it 'Failed to create a Robolectric sandbox: Android SDK 29
required Java 9 (have Java 8)'. Thus, in order to resolve this issue
and be able to run Robolectric tests from Android Studio, the
'Android JUnit' configuration need to be updated to point to 'Java 9+'
from the previous 'Default', which was taken from the Android
installation. The exact same need to happen for the command line,
running 'java --version' should point to 'Java 9+' in order for
Robolectric to be able to work.