[movies] build: Update dependency version naming to show stable versions
Just as with what was done for 'Kotlin' and 'Kotlin EAP', this change
makes sure that the rest of the dependencies that are using non stable
versions are being showcased as easy as possible. Also, as part of this
change, their corresponding stable versions are also being added for
better reference if need be.