[movies] update: Update android studio to canary 4.2.0-alpha04 version
Also, as part of this update the 'androidx.lifecycle:lifecycle-livedata-
core-ktx' dependency was removed from the 'core' module since it is now
marked as unused by the 'dependency-analysis' plugin.

This has probably happen due to the fact that this transient dependency
was wrongly added in the first place in the past. Or this has to do with
AndroidX and Jetifier playing some games in this release of AGP.