[movies-26] feature: Add to date extension function for date class
This will be used later on within the 'database' module and utilized by
the 'MovieEntity' so that a 'Movie' to a 'MovieEntity' before adding
the movie to the 'Movies.db' database. And since the 'Date' object needs
a bit of manipulation before it can be successfully added as a database
field, it is converted to a 'String' object and then being inserted.

Also, update the 'coverage.properties' file accordingly.