[movies-27] feature: Update movie entity to include paging fields
These pagination related fields will help later on to provide pagination
support to the 'RemoteMediator', which will query these values in order
to determine which page to load next, or if it should do that at all.

In more detail:
- The new autogenerated primary key 'id' field is required because the
previous 'movieId' primary key is not ordered appropriately. This is
because when the network response comes back, the list of movies is not
ordered by its 'movieId' (as expected), but that causes the pagination
library to end up shorting the by 'movieId' list itself, which will
result in the wrong previous and next page per item, when that item is
will be queried.
- The new 'prevPage' and 'nextPage' fields will hold the necessary
information, per item, that will help the 'RemoteMediator' decide
whether or not to load the next/previous page, and more specifically,
which page to load for the respective load type (refresh, append,
prepend).

Note that it was by design that those fields are being included within
the 'movies' table, instead of creating another separate table. This was
done in order to reduce complexity between table relations and database
calls. In the future and if need be, the 'movies' table will be split
accordingly.