[movies-27] feature: Add new load movies stream function to movies repo
This new paging related 'loadMoviesStream(...)' function will eventually
replace its above 'loadMovies(...)' function and provide to its clients
new paging features, through the new paging library.

The only problem is the fact that now the 'domain' module needs to
depend on an Android library, which is not the ideal. However, this is
unavoidable since the 'MoviesRepository' needs to know the type of the
'PagingData' class.