[movies-27] feature: Enhance load movie use case with load movies stream
Also, as part of this change a new way of invoking the use case was
introduced. This is by using operator overloading and utilising the
'invoke' function so that a client can use the use case like below:
- Instead of: 'useCase.execute(params)'
- To be doing: 'useCase(params)'
This makes '.execute' redundant and thus the usage of a use case much
simpler.