[movies-27] feature: Replace load movies with load movies stream
This change is about replacing the 'loadMoviesUseCase.execute(...)',
which triggers 'loadMovies(...)', with the newest
'loadMoviesUseCase(...)' trigger, which trigger 'loadMoviesStream(...)',
which utilises paging library.

As part of this change the below occurred:
1) The 'renderStatus(...)' function at the fragment level is simplified
due to the 'MoviesPagingAdapter' change, which does not require a manual
status update. All that are now being handled automatically by the
paging library.
2) The 'renderMovies(...)' function at the fragment level was replaced
with a similar function, but which accepts a parameter type of
'PagingData<Movie>' instead of the whole 'MoviesState'. This is an
improvement since the rest of the complexity is no longer required when
using the new 'MoviesPagingAdapter'.
3) The movies 'PaginationData<Movie>' state type was replaced with the
new 'PagingData<Movie>' type since this is what is now returned by the
use case.
4) The 'loadMovies(...)' view model function now collects the flow from
the use case 'loadMoviesStream', and instead of checking for success
and error it now just reacts on success only. The error handling will be
coming from the 'MoviesPagingAdapter' itself and its solution will be
introduced in a separate commit.
5) All test have been updated to reflect the above changes, but due to
the nature of the 'PagingData' class, which is not a data class like its
counterpart 'PaginationData' class, not all test can be replicated, as
such some of them are ignored, while others are just working for
demonstrative diffing purposes only and do not provide actual value.