[movies-27] feature: Add load state listener to handle error load states
As described in one of the previous commits, using the paging library,
the error handling is now originated from the 'PagingAdapter' instead of
this being done as a result of a use case call. This commit is all about
propagating the error trigger to view model, which will in turn update
the state to trigger a ui event.