[movies-27] cleanup: Remove movies status from movies state
'MovieStatus' is no longer required and as such it can be totally
removed from the 'MoviesState' logic. This status is being automatically
handled by the paging library itself.

Also, as part of this cleanup, 'IdleMovies' was removed and 'LoadMovies'
replaced this 'onResume' functionality, since previously the
'renderStatus(...)' was responsible for calling the 'LoadMovies', but
not it has to be done explicitly.

As such, the code become much more straightforward yet again.