[movies-27] cleanup: Remove movies adapter and related code
As part of this clean the whole 'core' module 'list' package and its
related files were removed as well. All those are no longer required
since the paging library will be now doing all this work instead.

Also, and since the list related code was the only code that was being
testing within the 'core' module, now that this is gone, the whole
testing section is gone with it. As such, the 'core' module becomes a
modules that does not require testing.