[movies-26] feature: Add new load date use case
This use case will help with the offline support. It will be responsible
for loading the date (year and month) from the database when the app is
launch while offline. This is required since might close the app while
having a year and/or month filter. As such, the list of movies persisted
in the database would be of that year and month. Thus, when the user
reopens the app while offline, they will be seeing a list of movies for
that specific date. This use case will provide support for that so that
the correct year and month is presented to the user when that happens.