[movies-27] test: Update instrumentation tests
Due to the change, from manual pagination to using the paging library
for that matter, an api call to load more movies will not be triggered
twice, loading both, the first and second pages during an initial load.
As such both instrumentation tests were update to expect two pages
instead of one.

Also, due to another change, loading a movie from the database instead
of network, when the users clicks on a specific movie, there is no need
for mocking those api calls. As such, all those were removed.