[movies-27] fix: Cancel load movies job before triggering a new load
This is done because of the fact that 'loadMovies(...)' is a flow and
can be triggered multiple times through 'reloadMovies(...)'. As such,
the previous job needs to be cancelled before a new load is triggered.
Otherwise, 'onLoadMoviesSuccess' will be triggered multiple times, from
multiple jobs, which will make the view model leak in a sense.

Note that this is not required with 'loadDate()' since it is not a flow,
nor it can be triggered multiple times.

The same applies to the 'loadMovie(...)' within 'MovieDetailsViewModel'.
Although it is a flow, it is not required that it is cancelled since it
will not get triggered multiple times.