[movies-27] test: Update coverage based on recent changes
Since paging library and offline support was added to the project, a lot
of changes occurred. Some of these changes caused a general drop of
coverage on the 'domain', 'network', 'data', 'movies' and
'movie-details' modules since the 'PagingData' related code cannot be
easily tested and as such has been left untested.