[movies-27] build: Add transitively used dependencies for data module
This change was advised by the dependency analysis plugin.

Also as part of this change and since the 'Room Runtime' and the
'Room KTX' dependencies were explicitly added to the 'data' module, it
is no longer required that those are defined as 'api' within the
'database' module. As such this is now reverted to 'implementation'.