[movies] build: Add explicit kotlin dependency on the build src module
This is done because after updating to Gradle distribution 6.6-rc-3 a
lot of files within the 'buildScr' module, like the 'Config' file, were
showing compilation errors due to for example not being able to compile
the 'listOf(...)' kotlin function. By explicitly adding the kotlin
dependency this issue is now fixed.

Not that the 'Repos', 'Plugins' and 'Deps' structure was copy-pasted as
is from the respective Kotlin files from within the module, since it
cannot work otherwise. As such, when the Kotlin dependency is updated,
it should get also update within the 'buildSrc' module as well.