[movies] fix: Remove initial load size manual configuration
This is done because as it is right now, because of the
'initialLoadSize' being overwritten with 'MOVIES_PAGE_SIZE', during a
refresh, 2 pages are being loaded instead of the new default of 4 pages.
However, this cause problems to the UI where it delays in loading new
pages when scrolling, which didn't happen with the previous alpha
version of paging 3 library. However, by moving back to the default of
loading 4 pages, the UI seems to be working much better. As such this
commits reverts the change to the default.