[movies] test: Add a delay while waiting for initial load of pages
The instrumentation test were starting to fail quite often after
updating to the new alpha paging release. Also, after changing the
initial load of pages from 2 to 4 pages, the tests started to fail more
often. As such a bit of delay was added for every initial load of pages
so that the app waits at least 1 second before checking on movies.

The above fixes some of the problems, but unfortunately not all of them.
The paging library sometimes skips some pages due to paging library
related errors, as such when that happens the instrumentation tests fail
since they expect a page to be loaded, but another page was loaded
instead. One of the error that occurs is the below:
'kotlinx.coroutines.JobCancellationException: ProducerCoroutine was
cancelled; job=ProducerCoroutine{Cancelling}@453f1f2'