[movies] fix: Comment out strict mode detect disk reads due to stateful
The stateful library is causing disk reads on startup, which is causing
strict mode to complain about them and throw a couple of dialogs for the
dev user to click 'OK' before continuing with the app. This is not the
ideal dev user experience and thus it is being commented out to improve
that. However, the detection of disk reads will no longer work until
this is fixed.