[movies] analysis: Change detekt auto correct to true
Detekt itself provides a wrapper over KtLint as a formatting rule set.

This change works hand-by-hand with the
'detektPlugins(Plugins.DETEKT_FORMATTING)' module configuration. when
that is present on a module and 'autoCorrect = true' is set then this
allows detekt to auto format the code upon running the 'detekt' task.