[movies] update: Update leak canary to latest 2.5 version
However, this update introduces a strict mode violation when it is
started for the first time. The dialog show is the below:

"The app Movies (process io.petros.movies has violated its
self-enforced StrictMode policy"

The logcat error message is the below:

"D/StrictMode: StrictMode policy violation; ~duration=3 ms:
android.os.strictmode.DiskWriteViolation"

This probably will be fixed with a new version of LeakCanary, as such
it should be closely monitored.