[movies] update: Update koin to latest 2.2.0 version
As part of this update the below occurred:
1) The Koin 'AndroidLogger()' function works as expect, the bug fix has
been implemented. As such it is no longer necessary to specify a level
of 'error' in order to have this working, any level now works as
expected. The default is 'info'.
2) A new 'saved state' beta dependency was added to the project and
feature module since those are transitively used by Koin and the
dependency analysis plubgin complains about it.