[movies] update: Update android test deps to latest alpha versions
1) Update android test core to latest 1.4.0-alpha04 version
2) Update android test espresso to latest 3.4.0-alpha04 version
3) Update android test junit to latest 1.1.3-alpha04 version

This update in necessary because apps targeting Android 12 and higher
are required to specify an explicit value for `android:exported` when
the corresponding component has an intent filter defined. However, the
current stable android test deps don't match this requirement.

Also, as part of this update the below occurred:
- 'org.hamcrest:hamcrest-all' was added to android test dependencies,
due to the fact that the alpha Espress dependencies don't include the
'Matchers' class which is required even for Espresso own implementation.
- 'org.hamcrest' group was excluded from the 'app' module Espresso
dependencies, due to the duplicate class problem with the
'hamcrest-integration' dependency.