[movies] update: Update leak canary to latest 2.7 version
This LeakCanary update fixes the 'android:exported' issue with apps
targeting Android 12. Also, it fixes the pending intent issue with
'FLAG_IMMUTABLE'. However, the latest Android S emulator is still
pointing to 30 for 'Build.VERSION.SDK_INT'. As such, when LeakCanary
tries to launch the notification it bypasses setting the
'FLAG_IMMUTABLE' (see 'NotificationReceiver' line 44), which causes the
app to crash with the bellow:

'java.lang.IllegalArgumentException: io.petros.movies: Targeting S+
(version 10000 and above) requires that one of FLAG_IMMUTABLE or
FLAG_MUTABLE be specified when creating a PendingIntent. Strongly
consider using FLAG_IMMUTABLE, only use FLAG_MUTABLE if some
functionality depends on the PendingIntent being mutable, e.g. if it
needs to be used with inline replies or bubbles.'

So, until the API S emulators point to the correct 'Build.VERSION
.SDK_INT' version, or you are using Android 12 preview on a real device,
LeakCanary cannot be utilised.