[movies] analysis: Remove test and multiplat. related detekt exclusions
This new configuration will make Detekt work on the test level as well
as it does on the main code level, with a few exceptions.

Also, since this project is not yet multiplatform, those exclusions are
removed as well, just to make it more explicit that nothing like that is
currently utilised within this project.