[movies-29] feature: Connect compose with the movie details screen
This change effectively enables switching between 'Jetpack Compose' and
the 'Legacy Android View' implementations of the 'Movie Details' screen.

A user will be able to enable/disable 'Jetpack Compose' from the drawer
and this way, to switch between the two implementations.

Additionally, the user will be able to tell which implementation is
which, through the toolbar, which will be having the 'Compose' suffix
added on the title when 'Jetpack Compose' is enabled.