[movies] update: Update dependency analysis plugin to latest 0.75.0
This update fixes the 'gradlew buildHealth' task, which started
failing with the below exception after upgrading gradle android plugin
to 7.0.0-alpha15 version.

* What went wrong:
Execution failed for task ':app:findAndroidLintersDebug'.
> java.nio.file.NoSuchFileException:
...\core\build\intermediates\lint_publish_jar\global\lint.jar

Also, as part of this update and because the plugin was working for a
bit of time now, few dependencies and configurations needed a bit of
tuning to keep the build as stable as possible, based on the plugin's
advise.