[movies] lib: Remove bye-bye jetifier gradle plugin
This gradle plugin to verify if a project can keep Android Jetifier or
not is no longer needed and thus can be removed from the project.

All current and most probably future libraries that are to be added to
this project are/will not be having jetifier in them. Thus, this
now makes this plugin redundant.